feat: do changes for code review
