+++ apps/eo_web/src/api/useElixirApi.ts
@@ -1,32 +1,6 @@
 import { api } from "~/api/axios";
 import { API_URL } from "~/api/common";
-import { useProfileStore } from "~/stores/useProfileStore";
-
-export interface ZipCodeValidationResponseSuccess {
-  app_flags: {
-    can_delete_sessions: boolean;
-    can_edit_sessions: boolean;
-    can_mute_sessions: boolean;
-    malady_selection: Array<{
-      label: string;
-      slug: string;
-    }>;
-  };
-  birth_date: string | null;
-  caregiver: boolean;
-  ced: boolean;
-  email: string;
-  first_name: string | null;
-  gender: string | null;
-  last_name: string | null;
-  malady: string;
-  med_card: boolean;
-  phone: unknown;
-  status: string;
-  timezone: string;
-  uid: string;
-  zip: string;
-}
+import { useProfileStore, type Profile } from "~/stores/useProfileStore";
 
 export interface ZipCodeValidationResponseError {
   errors: {
@@ -49,9 +23,7 @@ export const useElixirApi = () => {
     },
   };
   const validateZipCode = async (zipCode: string) => {
-    return api.post<
-      ZipCodeValidationResponseSuccess | ZipCodeValidationResponseError
-    >(
+    return api.post<Profile | ZipCodeValidationResponseError>(
       `${API_URL}/v2/profile/validate_zip_code`,
       {
         zip: zipCode,
